+++ src/main/java/TallerReparacionesSpringBoot/entities/Usuario.java
@@ -11,7 +11,7 @@ import jakarta.persistence.Id;
 public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_usuario;
+    private Long id_usuario;
     
     private String nombre_usuario;
     private String dni_usuario;
@@ -32,11 +32,11 @@ public class Usuario {
     }
     
     // Getters y Setters
-    public int getId_usuario() {
+    public Long getId_usuario() {
         return id_usuario;
     }
     
-    public void setId_usuario(int id_usuario) {
+    public void setId_usuario(Long id_usuario) {
         this.id_usuario = id_usuario;
     }
